Move regression tests for useSnapshot perf improvement to optimization test file
